Activate a specific monitor using = in d
I wanted to be able to go to a specific monitor using a hotkey. If I'm already on that monitor, I don't want to move away. This change allows this by putting =<n> in the argument for the Manager_activateMonitor function. I've also made the change for Manager_setWindowMonitor just to be consistent.