Add Monitor_greedyView()
This is a hack to emulate XMonad's greedyView function.

It is similar to Monitor_activateView() but it ensures that there is
only ever one copy of a view (instead of one copy per monitor).  When we
switch to view X using Monitor_greedyView(), all windows from view X on
all monitors are moved to view X on the active monitor.

If the view we're switching to is currently visible on another monitor,
they are swapped.